........
  r302 | erno | 2008-02-28 12:28:12 +0200 (Thu, 28 Feb 2008) | 1 line

  metaformedit dies again
........
  r304 | therauli | 2008-02-28 13:32:45 +0200 (Thu, 28 Feb 2008) | 2 lines

  Fixed unicode bug in DeletePage
........
  r305 | therauli | 2008-02-28 13:49:17 +0200 (Thu, 28 Feb 2008) | 2 lines

  Fixed fixes for DeletePage
........
  r306 | therauli | 2008-02-28 14:20:58 +0200 (Thu, 28 Feb 2008) | 2 lines

  Fixed fixes in fixed Deletepage might work now
........
  r309 | therauli | 2008-02-28 15:08:05 +0200 (Thu, 28 Feb 2008) | 2 lines

  Fixed (hopefully) all bugs in DeletePage that were fixed before. DeletePage now throws an exception if the page does not exist.
........
  r310 | jviide | 2008-02-28 15:14:51 +0200 (Thu, 28 Feb 2008) | 3 lines

  Added a xmlrpc plugin for downloading files in several pieces. Uploading
  in pieces upcoming.
........
  r311 | oherrala | 2008-02-28 15:35:43 +0200 (Thu, 28 Feb 2008) | 2 lines

  Add preliminary unit test framework testing things over XML-RPC.
........
  r313 | oherrala | 2008-02-28 15:44:59 +0200 (Thu, 28 Feb 2008) | 2 lines

  Add boring GetMeta tests requested by someone at this Hackathon.
........

--HG--
branch : moin-1.6-branch

+++ moin/graphingwiki/plugin/xmlrpc/DeletePage.py
@@ -9,7 +9,9 @@
 
 import os
 import xmlrpclib
+
 from MoinMoin.PageEditor import PageEditor
+from MoinMoin import config
 
 def delete(request, pagename, comment = None):
     _ = request.getText
@@ -23,8 +25,12 @@ def delete(request, pagename, comment = None):
     if not request.user.may.delete(pagename):
         return xmlrpclib.Fault(1, _("You are not allowed to delete this page"))
     
-    #Deletespages
+    #Deletes pages
+
     page = PageEditor(request, pagename, do_editor_backup=0)
+    if not page.exists():
+        return xmlrpclib.Fault(1, _('No such page %s' % pagename))
+
     page.deletePage(comment)
 
     return True
@@ -35,8 +41,12 @@ def execute(xmlrpcobj, pagename, comment = None):
 
     pagename = xmlrpcobj._instr(pagename)
 
-    success = delete(request, pagename, comment)
+    if comment:
+        comment = xmlrpcobj._instr(comment)
+    else:
+        comment = u''
 
+    success = delete(request, pagename, comment)
 
     # Save, delete return True on success
     if success is True:
